Independent AnimatedFeederInfosStyleProvider
