+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/svg/styles/iidm/HighlightLineStateStyleProvider.java
@@ -10,18 +10,12 @@ package com.powsybl.sld.svg.styles.iidm;
 import com.powsybl.iidm.network.Branch;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.ThreeWindingsTransformer;
-import com.powsybl.sld.library.ComponentLibrary;
 import com.powsybl.sld.library.ComponentTypeName;
-import com.powsybl.sld.model.cells.Cell;
 import com.powsybl.sld.model.graphs.Graph;
-import com.powsybl.sld.model.graphs.VoltageLevelGraph;
 import com.powsybl.sld.model.nodes.*;
 import com.powsybl.sld.model.nodes.feeders.FeederWithSides;
-import com.powsybl.sld.svg.BusInfo;
-import com.powsybl.sld.svg.DiagramLabelProvider;
-import com.powsybl.sld.svg.FeederInfo;
+import com.powsybl.sld.svg.styles.EmptyStyleProvider;
 import com.powsybl.sld.svg.styles.StyleClassConstants;
-import com.powsybl.sld.svg.styles.StyleProvider;
 
 import java.util.*;
 
@@ -29,7 +23,7 @@ import java.util.*;
  * @author Franck Lecuyer <franck.lecuyer at rte-france.com>
  * @author Florian Dupuy <florian.dupuy at rte-france.com>
  */
-public class HighlightLineStateStyleProvider implements StyleProvider {
+public class HighlightLineStateStyleProvider extends EmptyStyleProvider {
 
     private final Network network;
 
@@ -133,52 +127,7 @@ public class HighlightLineStateStyleProvider implements StyleProvider {
     }
 
     @Override
-    public List<String> getSvgNodeStyles(VoltageLevelGraph graph, Node node, ComponentLibrary componentLibrary, boolean showInternalNodes) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<String> getSvgNodeDecoratorStyles(DiagramLabelProvider.NodeDecorator nodeDecorator, Node node, ComponentLibrary componentLibrary) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<String> getZoneLineStyles(BranchEdge edge, ComponentLibrary componentLibrary) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<String> getSvgNodeSubcomponentStyles(Graph graph, Node node, String subComponentName) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public void reset() {
-        // Nothing to do
-    }
-
-    @Override
     public List<String> getCssFilenames() {
         return List.of("highlightLineStates.css");
-    }
-
-    @Override
-    public List<String> getBusStyles(String busId, VoltageLevelGraph graph) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<String> getBusInfoStyle(BusInfo info) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<String> getFeederInfoStyles(FeederInfo info) {
-        return Collections.emptyList();
-    }
-
-    @Override
-    public List<String> getCellStyles(Cell cell) {
-        return Collections.emptyList();
     }
 }
